Implement color tool
Also, turns out that OpenEXR will store the channels in the file
alphabetically based on the channel names, so if you want to read back
into your structure based on the channel names, you have to name the
channels in such a manner that when sorted by name, they correspond to
the field order in your struct.  Wheee, aren't libraries fun?